[major] Add new `--from` and `--add` flags for syncing access

+++ command/access/index.js
@@ -40,7 +40,9 @@ var Clone = module.exports = Uni.Command.extend({
    * @public
    */
   flags: {
-    '--team': 'Don\'t sync the repo collaborators but sync an organization team'
+    '--team': 'Don\'t sync the repo collaborators but sync an organization team',
+    '--add': 'Don\'t add collaborators, a team but an individual account',
+    '--from': 'Get module names from an npm account instead of GitHub'
   },
 
   steps: {
@@ -52,6 +54,26 @@ var Clone = module.exports = Uni.Command.extend({
       var github = this.githulk.project(this.uni.argv[0])
         , cmd = this;
 
+      //
+      // uni access --from V1 --add v1
+      //
+      if (this.uni.flag.from) {
+        return this.registry.users.list(this.uni.flag.from, function list(err, names) {
+          if (err) return next(err);
+
+          names.forEach(function each(pkg) {
+            cmd.packages[pkg.name] = {
+              name: pkg.name
+            };
+          });
+
+          return next();
+        });
+      }
+
+      //
+      // uni access foo/bar npmname
+      //
       if (this.uni.argv[1]) {
         this.packages[github.project || this.uni.argv[1]] = {
           name: this.uni.argv[1]
@@ -60,6 +82,9 @@ var Clone = module.exports = Uni.Command.extend({
         return next();
       }
 
+      //
+      // No known packages that should be synced.
+      //
       this.projects(github, function list(err, projects) {
         if (err) return next(err);
 
@@ -88,6 +113,18 @@ var Clone = module.exports = Uni.Command.extend({
     collaborators: function collaborators(next) {
       var github = this.githulk.project(this.uni.argv[0])
         , cmd = this;
+
+      //
+      // We only have to sync one user account, not multiple so we can add them
+      // straight away
+      //
+      if (this.uni.flag.add) {
+        Object.keys(cmd.packages).forEach(function each(name) {
+          cmd.packages[name].users = [cmd.uni.flag.add];
+        });
+
+        return next();
+      }
 
       //
       // We either need to sync all the collaborators of the github repository
@@ -129,6 +166,11 @@ var Clone = module.exports = Uni.Command.extend({
       var confirmed = {}
         , cmd = this;
 
+      //
+      // Manual entry, assume name is already valid.
+      //
+      if (this.uni.flag.add) return next();
+
       async.eachSeries(Object.keys(this.packages), function each(name, next) {
         var users = cmd.packages[name].users;
 
@@ -148,11 +190,13 @@ var Clone = module.exports = Uni.Command.extend({
               message: 'If '+ github +' is not '+ npm +' on npm, supply the correct username'
             }], function answered(answer) {
               confirmed[github] = answer.name.trim() || npm;
+
               return fn(undefined, confirmed[github]);
             });
           });
         }, function complete(err, users) {
-          cmd.users = users;
+          cmd.packages[name].users = users;
+
           next();
         });
       });
@@ -169,7 +213,14 @@ var Clone = module.exports = Uni.Command.extend({
       async.eachSeries(Object.keys(this.packages), function each(name, next) {
         var pkg = cmd.packages[name];
 
-        cmd.maintainersDiff(pkg.name, pkg.users, function diff(err, stat) {
+        /**
+         * The diff that we need to add.
+         *
+         * @param {Error} err Failed operation.
+         * @param {Object} stat Diff result contain add/remove array.
+         * @api public
+         */
+        function diff(err, stat) {
           if (err) return next(err);
 
           //
@@ -179,13 +230,13 @@ var Clone = module.exports = Uni.Command.extend({
           var me = uni.conf.get('username');
 
           stat.add.forEach(function add(name) {
-            console.log('adding', name, 'as maintainer of', pkg.name);
-            npm().owner('add '+ name +' '+ pkg.name);
+            console.log('adding %s as maintainer of %s', name, pkg.name);
+            if (!uni.flag.dryrun) npm().owner('add '+ name +' '+ pkg.name);
           });
 
           stat.remove.forEach(function add(name) {
-            console.log('removing', name, 'as maintainer of', pkg.name);
-            npm().owner('rm '+ name +' '+ pkg.name);
+            console.log('removing %s as maintainer of %s', name, pkg.name);
+            if (!uni.flag.dryrun) npm().owner('rm '+ name +' '+ pkg.name);
           });
 
           if (!stat.remove.length && !stat.add.length) {
@@ -193,7 +244,10 @@ var Clone = module.exports = Uni.Command.extend({
           }
 
           next();
-        });
+        }
+
+        if (uni.flag.add) return diff(undefined, { add: pkg.users, remove: [] });
+        return cmd.maintainersDiff(pkg.name, pkg.users, diff);
       }, next);
     }
   },
